+++ djangosecure/conf.py
@@ -21,6 +21,7 @@ conf = Configuration(
     SECURE_HSTS_SECONDS=0,
     SECURE_FRAME_DENY=False,
     SECURE_CONTENT_TYPE_NOSNIFF=False,
+    SECURE_XSS_FILTER=False,
     SECURE_SSL_REDIRECT=False,
     SECURE_SSL_HOST=None,
     SECURE_REDIRECT_EXEMPT=[],
@@ -33,6 +34,7 @@ conf = Configuration(
         "djangosecure.check.djangosecure.check_sts",
         "djangosecure.check.djangosecure.check_frame_deny",
         "djangosecure.check.djangosecure.check_content_type_nosniff",
+        "djangosecure.check.djangosecure.check_xss_filter",
         "djangosecure.check.djangosecure.check_ssl_redirect",
         ]
     )