+++ djangosecure/middleware.py
@@ -10,6 +10,7 @@ class SecurityMiddleware(object):
         self.sts_seconds = conf.SECURE_HSTS_SECONDS
         self.frame_deny = conf.SECURE_FRAME_DENY
         self.content_type_nosniff = conf.SECURE_CONTENT_TYPE_NOSNIFF
+        self.xss_filter = conf.SECURE_XSS_FILTER
         self.redirect = conf.SECURE_SSL_REDIRECT
         self.redirect_host = conf.SECURE_SSL_HOST
         self.proxy_ssl_header = conf.SECURE_PROXY_SSL_HEADER
@@ -40,13 +41,18 @@ class SecurityMiddleware(object):
                 not getattr(response, "_frame_deny_exempt", False) and
                 not 'x-frame-options' in response):
             response["x-frame-options"] = "DENY"
+
         if (self.sts_seconds and
                 request.is_secure() and
                 not 'strict-transport-security' in response):
             response["strict-transport-security"] = ("max-age=%s"
                                                      % self.sts_seconds)
+
         if (self.content_type_nosniff and
                 not 'x-content-type-options' in response):
             response["x-content-type-options"] = "nosniff"
+
+        if self.xss_filter and not 'x-xss-protection' in response:
+            response["x-xss-protection"] = "1; mode=block"
 
         return response